requirements.txt
flask>=3.0.0
line-bot-sdk>=3.0.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
